refactor(profile page): componentize, clean up code
