+++ src/components/profile/user/index.ts
@@ -1,0 +1,4 @@
+export { default as UserHeaderBox, getOrdinalSuffix } from "./UserHeaderBox";
+export { default as UserStatsBox } from "./UserStatsBox";
+export { default as TensaraScoreBox } from "./TensaraScoreBox";
+export { default as UserLanguagesBox } from "./UserLanguagesBox";
